+++ src/lib.rs
@@ -2,7 +2,7 @@
 #![allow(clippy::needless_doctest_main)]
 #![deny(unsafe_op_in_unsafe_fn)]
 #![warn(missing_docs)]
-#![cfg_attr(docsrs, feature(doc_auto_cfg))]
+#![cfg_attr(docsrs, feature(doc_cfg))]
 
 extern crate core;